PR # 2

+++ src/main/java/com/auction/service/AuctionServiceImpl.java
@@ -1,13 +1,14 @@
 package com.auction.service;
-
 import com.auction.dao.AuctionRepository;
 import com.auction.dto.AuctionDTO;
 import com.auction.entities.Auction;
 import com.auction.entities.Product;
-import com.auction.exception.ResourceNotFoundException;
+import com.auction.exception.AuctionNotFoundException;
+import com.auction.exception.AuctionStartDatePassedException;
+import com.auction.exception.DateValidationException;
+import com.auction.exception.ProductNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.time.ZoneId;
 import java.util.Date;
 import java.util.List;
@@ -29,14 +30,18 @@ public class AuctionServiceImpl implements AuctionService {
     @Override
     public Auction findById(int auctionId) {
         return auctionRepository.findById(auctionId)
-                .orElseThrow(() -> new ResourceNotFoundException("Auction not found with id - " + auctionId));
+                .orElseThrow(() -> new AuctionNotFoundException(auctionId));
     }
 
     @Override
     public Auction createAuction(int productId, AuctionDTO auctionDTO) {
         Product product = productService.findById(productId);
         if (product == null) {
-            throw new ResourceNotFoundException("Product not found with id - " + productId);
+            throw new ProductNotFoundException(productId);
+        }
+
+        if (!auctionDTO.getEndDate().isAfter(auctionDTO.getStartDate())) {
+            throw new DateValidationException("End date must be after start date.");
         }
 
         Auction newAuction = new Auction();
@@ -47,10 +52,14 @@ public class AuctionServiceImpl implements AuctionService {
 
         return auctionRepository.save(newAuction);
     }
-
     @Override
     public void deleteById(int auctionId) {
         Auction auction = findById(auctionId);
+
+        Date now = new Date();
+        if (now.after(auction.getStartDate())) {
+            throw new AuctionStartDatePassedException("Cannot delete the auction. The start date has already passed.");
+        }
         auctionRepository.deleteById(auctionId);
     }
 }
